Remove magic number 32s from the game and level logic Entities can be different sizes than just 1x1 Add constructor for SimpleVector on the lua side In DEBUG builds, show a magenta square for the entity size and positions

+++ src/Game/Game.h
@@ -37,7 +37,7 @@ private:
 	WorldDefinition m_world;
 	LuaRegistry m_luaBindings;
 
-	uint64_t m_playerEntityID;
+	uint32_t m_playerEntityID;
 
 	const PortalTrigger* m_lastEnteredPortal;
